requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
numpy==1.25.2
Pyro4==4.82
python-dotenv==1.0.0
requests==2.31.0
serpent==1.41
urllib3==2.0.4
